requirements.txt
# Web Framework
streamlit>=1.28.0

# Data Analysis and Visualization
pandas>=2.0.0
yfinance>=0.2.28
plotly>=5.18.0
numpy>=1.24.0

# API and HTTP
requests>=2.31.0
python-dotenv>=1.0.0

# Nova/LLM Dependencies
nova-client>=1.0.0  # Replace with actual version if different

# Date and Time
python-dateutil>=2.8.2

# Optional: Development Tools
black>=23.9.1  # Code formatting
pylint>=2.17.0  # Code analysis